+++ Engine/Source/Runtime/Engine/Private/GameFramework/Components/HFirstPersonCameraComponent.cpp
@@ -21,19 +21,19 @@ void HFirstPersonCameraComponent::LookUp( float Value )
 {
 	const float kXRotationClamp = Math::DegreesToRadians( 88.f );
 
-	m_Rotation.x += Value * m_CameraPitchSpeedMultiplier * (float)GEngine->GetDeltaTime();
+	m_Rotation.x += Value * m_CameraPitchSpeedMultiplier * GEngine->GetDeltaTime();
 	m_Rotation.x = Math::Clamp( m_Rotation.x, -kXRotationClamp, kXRotationClamp );
 	SetRotation( FQuat::CreateFromYawPitchRoll( m_Rotation.y, m_Rotation.x, m_Rotation.z ) );
 }
 
 void HFirstPersonCameraComponent::LookRight( float Value )
 {
-	m_Rotation.y += Value * m_CameraYawSpeedMultiplier * (float)GEngine->GetDeltaTime();
+	m_Rotation.y += Value * m_CameraYawSpeedMultiplier * GEngine->GetDeltaTime();
 	SetRotation( FQuat::CreateFromYawPitchRoll( m_Rotation.y, m_Rotation.x, m_Rotation.z ) );
 }
 
 void HFirstPersonCameraComponent::Roll( float Value )
 {
-	m_Rotation.z += Value * m_CameraRollSpeedMultiplier * (float)GEngine->GetDeltaTime();
+	m_Rotation.z += Value * m_CameraRollSpeedMultiplier * GEngine->GetDeltaTime();
 	SetRotation( FQuat::CreateFromYawPitchRoll( m_Rotation.y, m_Rotation.x, m_Rotation.z ) );
 }
